Refactored some read only methods into own class.
