+++ orcid-core/src/main/java/org/orcid/core/manager/impl/OrcidProfileManagerImpl.java
@@ -49,21 +49,15 @@ import net.sf.ehcache.Element;
 
 import org.apache.commons.lang3.StringUtils;
 import org.orcid.core.adapter.Jaxb2JpaAdapter;
-import org.orcid.core.adapter.Jpa2JaxbAdapter;
-import org.orcid.core.adapter.JpaJaxbEntityAdapter;
 import org.orcid.core.constants.DefaultPreferences;
-import org.orcid.core.locale.LocaleManager;
-import org.orcid.core.manager.EncryptionManager;
 import org.orcid.core.manager.LoadOptions;
 import org.orcid.core.manager.NotificationManager;
 import org.orcid.core.manager.OrcidGenerationManager;
 import org.orcid.core.manager.OrcidIndexManager;
 import org.orcid.core.manager.OrcidJaxbCopyManager;
-import org.orcid.core.manager.OrcidProfileCacheManager;
 import org.orcid.core.manager.OrcidProfileCleaner;
 import org.orcid.core.manager.OrcidProfileManager;
 import org.orcid.core.manager.OrgManager;
-import org.orcid.core.manager.SourceManager;
 import org.orcid.core.security.OrcidWebRole;
 import org.orcid.core.security.visibility.OrcidVisibilityDefaults;
 import org.orcid.core.security.visibility.aop.VisibilityControl;
@@ -71,7 +65,6 @@ import org.orcid.jaxb.model.message.ActivitiesVisibilityDefault;
 import org.orcid.jaxb.model.message.Affiliation;
 import org.orcid.jaxb.model.message.Affiliations;
 import org.orcid.jaxb.model.message.Biography;
-import org.orcid.jaxb.model.message.Claimed;
 import org.orcid.jaxb.model.message.ContactDetails;
 import org.orcid.jaxb.model.message.Contributor;
 import org.orcid.jaxb.model.message.ContributorOrcid;
@@ -87,12 +80,9 @@ import org.orcid.jaxb.model.message.FamilyName;
 import org.orcid.jaxb.model.message.Funding;
 import org.orcid.jaxb.model.message.FundingList;
 import org.orcid.jaxb.model.message.GivenNames;
-import org.orcid.jaxb.model.message.LastModifiedDate;
 import org.orcid.jaxb.model.message.OrcidActivities;
 import org.orcid.jaxb.model.message.OrcidBio;
-import org.orcid.jaxb.model.message.OrcidDeprecated;
 import org.orcid.jaxb.model.message.OrcidHistory;
-import org.orcid.jaxb.model.message.OrcidIdentifier;
 import org.orcid.jaxb.model.message.OrcidInternal;
 import org.orcid.jaxb.model.message.OrcidPreferences;
 import org.orcid.jaxb.model.message.OrcidProfile;
@@ -139,15 +129,8 @@ import org.orcid.persistence.jpa.entities.SourceEntity;
 import org.orcid.persistence.jpa.entities.WorkEntity;
 import org.orcid.pojo.ajaxForm.PojoUtil;
 import org.orcid.utils.DateUtils;
-import org.orcid.utils.NullUtils;
 import org.orcid.utils.OrcidStringUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.transaction.TransactionStatus;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.support.TransactionCallbackWithoutResult;
@@ -167,16 +150,10 @@ import org.springframework.transaction.support.TransactionTemplate;
  * @author Declan Newman and Will Simpson
  */
 
-public class OrcidProfileManagerImpl implements OrcidProfileManager {
-
-    @Resource
-    private EncryptionManager encryptionManager;
+public class OrcidProfileManagerImpl extends OrcidProfileManagerReadOnlyImpl implements OrcidProfileManager {
 
     @Resource
     private OrcidGenerationManager orcidGenerationManager;
-
-    @Resource
-    private ProfileDao profileDao;
 
     @Resource
     private ProfileWorkDao profileWorkDao;
@@ -197,9 +174,6 @@ public class OrcidProfileManagerImpl implements OrcidProfileManager {
     private OrcidOauth2TokenDetailDao orcidOauth2TokenDetailDao;
 
     @Resource
-    private JpaJaxbEntityAdapter adapter;
-
-    @Resource
     private Jaxb2JpaAdapter jaxb2JpaAdapter;
 
     @Resource
@@ -215,16 +189,7 @@ public class OrcidProfileManagerImpl implements OrcidProfileManager {
     private OrcidProfileCleaner orcidProfileCleaner;
 
     @Resource
-    private LocaleManager localeManager;
-
-    @Resource
     private GenericDao<EmailEventEntity, Long> emailEventDao;
-
-    @Resource
-    private SourceManager sourceManager;
-
-    @Resource
-    private OrcidProfileCacheManager orcidProfileCacheManager;
 
     @Resource
     private OrcidJaxbCopyManager orcidJaxbCopyManager;
@@ -235,13 +200,8 @@ public class OrcidProfileManagerImpl implements OrcidProfileManager {
     @Resource
     private OrgManager orgManager;
     
-    @Resource
-    private Jpa2JaxbAdapter jpaJaxbAdapter;
-
     @Value("${org.orcid.core.works.compare.useScopusWay:false}")
     private boolean compareWorksUsingScopusWay;
-
-    private int claimWaitPeriodDays = 10;
 
     private int claimReminderAfterDays = 8;
 
@@ -257,8 +217,6 @@ public class OrcidProfileManagerImpl implements OrcidProfileManager {
 
     private int numberOfIndexingThreads;
 
-    private static final Logger LOG = LoggerFactory.getLogger(OrcidProfileManagerImpl.class);
-
     private static final int INDEXING_BATCH_SIZE = 100;
 
     public void setProfileDao(ProfileDao profileDao) {
@@ -271,10 +229,6 @@ public class OrcidProfileManagerImpl implements OrcidProfileManager {
 
     public void setNumberOfIndexingThreads(int numberOfIndexingThreads) {
         this.numberOfIndexingThreads = numberOfIndexingThreads;
-    }
-
-    public void setClaimWaitPeriodDays(int claimWaitPeriodDays) {
-        this.claimWaitPeriodDays = claimWaitPeriodDays;
     }
 
     public void setClaimReminderAfterDays(int claimReminderAfterDays) {
@@ -509,198 +463,6 @@ public class OrcidProfileManagerImpl implements OrcidProfileManager {
                 }
             }
         }
-    }
-
-    /**
-     * Retrieves the orcid external identifiers given an identifier
-     * 
-     * @param orcid
-     *            the identifier
-     * @return the orcid profile with only the bio populated
-     */
-    @Override
-    @Transactional
-    public OrcidProfile retrieveClaimedExternalIdentifiers(String orcid) {
-        OrcidProfile profile = retrieveClaimedOrcidProfile(orcid);
-        if (profile != null) {
-            profile.downgradeToExternalIdentifiersOnly();
-        }
-        return profile;
-    }
-
-    /**
-     * Retrieves the orcid bio given an identifier
-     * 
-     * @param orcid
-     *            the identifier
-     * @return the orcid profile with only the bio populated
-     */
-    @Override
-    @Transactional
-    public OrcidProfile retrieveClaimedOrcidBio(String orcid) {
-        OrcidProfile profile = retrieveClaimedOrcidProfile(orcid);
-        if (profile != null) {
-            profile.downgradeToBioOnly();
-        }
-        return profile;
-    }
-
-    /**
-     * Retrieves the orcid affiliations given an identifier
-     * 
-     * @param orcid
-     *            the identifier
-     * @return the orcid profile with only the affiliations populated
-     */
-    @Override
-    @Transactional
-    public OrcidProfile retrieveClaimedAffiliations(String orcid) {
-        OrcidProfile profile = retrieveClaimedOrcidProfile(orcid);
-        if (profile != null) {
-            profile.downgradeToAffiliationsOnly();
-        }
-        return profile;
-    }
-
-    /**
-     * Retrieves the orcid fundings given an identifier
-     * 
-     * @param orcid
-     *            the identifier
-     * @return the orcid profile with only the funding list populated
-     */
-    @Override
-    @Transactional
-    public OrcidProfile retrieveClaimedFundings(String orcid) {
-        OrcidProfile profile = retrieveClaimedOrcidProfile(orcid);
-        if (profile != null) {
-            profile.downgradeToFundingsOnly();
-        }
-        return profile;
-    }
-
-    /**
-     * Retrieves the orcid works given an identifier
-     * 
-     * @param orcid
-     *            the identifier
-     * @return the orcid profile with only the works populated
-     */
-    @Override
-    @Transactional
-    public OrcidProfile retrieveClaimedOrcidWorks(String orcid) {
-        OrcidProfile profile = retrieveClaimedOrcidProfile(orcid);
-        if (profile != null) {
-            profile.downgradeToWorksOnly();
-        }
-        return profile;
-    }
-
-    @Override
-    @Transactional
-    public OrcidProfile retrieveOrcidProfile(String orcid) {
-        return retrieveOrcidProfile(orcid, LoadOptions.ALL);
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public OrcidProfile retrieveOrcidProfile(String orcid, LoadOptions loadOptions) {
-        if (LoadOptions.ALL.equals(loadOptions))
-            return orcidProfileCacheManager.retrieve(orcid);
-        return retrieveFreshOrcidProfile(orcid, loadOptions);
-    }
-
-    @Transactional
-    public OrcidProfile retrieveFreshOrcidProfile(String orcid, LoadOptions loadOptions) {
-        LOG.debug("About to obtain fresh profile: " + orcid);
-        profileDao.flush();
-        ProfileEntity profileEntity = profileDao.find(orcid);
-        if (profileEntity != null) {
-            OrcidProfile freshOrcidProfile = convertToOrcidProfile(profileEntity, loadOptions);
-            return freshOrcidProfile;
-        }
-        return null;
-    }
-
-    @Override
-    @Transactional
-    public OrcidProfile retrieveClaimedOrcidProfile(String orcid) {
-        OrcidProfile orcidProfile = retrieveOrcidProfile(orcid);
-        if (orcidProfile != null) {
-            if (Boolean.TRUE.equals(orcidProfile.getOrcidHistory().getClaimed().isValue()) || orcidProfile.isDeactivated()
-                    || isBeingAccessedByCreator(orcidProfile) || haveSystemRole() || isOldEnough(orcidProfile)) {
-                return orcidProfile;
-            } else {
-                if (orcidProfile.getOrcidDeprecated() != null && orcidProfile.getOrcidDeprecated().getPrimaryRecord() != null)
-                    return createReservedForClaimOrcidProfile(orcid, orcidProfile.getOrcidDeprecated(), orcidProfile.getOrcidHistory().getLastModifiedDate());
-                else
-                    return createReservedForClaimOrcidProfile(orcid, orcidProfile.getOrcidHistory().getLastModifiedDate());
-            }
-        }
-        return null;
-    }
-
-    private boolean isOldEnough(OrcidProfile orcidProfile) {
-        return DateUtils.olderThan(orcidProfile.getOrcidHistory().getSubmissionDate().getValue().toGregorianCalendar().getTime(), claimWaitPeriodDays);
-    }
-
-    private boolean isBeingAccessedByCreator(OrcidProfile orcidProfile) {
-        String amenderOrcid = sourceManager.retrieveSourceOrcid();
-        Source source = orcidProfile.getOrcidHistory().getSource();
-        if (NullUtils.noneNull(amenderOrcid, source)) {
-            return amenderOrcid.equals(source.retrieveSourcePath());
-        }
-        return false;
-    }
-
-    private boolean haveSystemRole() {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        if (authentication != null) {
-            Collection<? extends GrantedAuthority> authorities = authentication.getAuthorities();
-            if (authorities != null) {
-                return authorities.contains(new SimpleGrantedAuthority("ROLE_SYSTEM"));
-            }
-        }
-        return false;
-    }
-
-    private OrcidProfile createReservedForClaimOrcidProfile(String orcid, LastModifiedDate lastModifiedDate) {
-        return createReservedForClaimOrcidProfile(orcid, null, lastModifiedDate);
-    }
-
-    private OrcidProfile createReservedForClaimOrcidProfile(String orcid, OrcidDeprecated deprecatedInfo, LastModifiedDate lastModifiedDate) {
-        OrcidProfile op = new OrcidProfile();
-        if(jpaJaxbAdapter != null) {
-        	 op.setOrcidIdentifier(new OrcidIdentifier(jpaJaxbAdapter.getOrcidIdBase(orcid)));
-        } else {
-        	op.setOrcidIdentifier(orcid);
-        }
-        if (deprecatedInfo != null)
-            op.setOrcidDeprecated(deprecatedInfo);
-
-        OrcidHistory oh = new OrcidHistory();
-        oh.setClaimed(new Claimed(false));
-        oh.setLastModifiedDate(lastModifiedDate);
-        op.setOrcidHistory(oh);
-        GivenNames gn = new GivenNames();
-        PersonalDetails pd = new PersonalDetails();
-        gn.setContent(localeManager.resolveMessage("orcid.reserved_for_claim"));
-        pd.setGivenNames(gn);
-        OrcidBio ob = new OrcidBio();
-        ob.setPersonalDetails(pd);
-        op.setOrcidBio(ob);
-        return op;
-    }
-
-    private OrcidProfile convertToOrcidProfile(ProfileEntity profileEntity, LoadOptions loadOptions) {
-        LOG.debug("About to convert profile entity to orcid profile: " + profileEntity.getId());
-        profileDao.refresh(profileEntity);
-        OrcidProfile orcidProfile = adapter.toOrcidProfile(profileEntity, loadOptions);
-        String verificationCode = profileEntity.getEncryptedVerificationCode();
-        String securityAnswer = profileEntity.getEncryptedSecurityAnswer();
-        orcidProfile.setVerificationCode(decrypt(verificationCode));
-        orcidProfile.setSecurityQuestionAnswer(decrypt(securityAnswer));
-        return orcidProfile;
     }
 
     @Override
@@ -2071,14 +1833,6 @@ public class OrcidProfileManagerImpl implements OrcidProfileManager {
     private String encrypt(String unencrypted) {
         if (StringUtils.isNotBlank(unencrypted)) {
             return encryptionManager.encryptForInternalUse(unencrypted);
-        } else {
-            return null;
-        }
-    }
-
-    private String decrypt(String encrypted) {
-        if (StringUtils.isNotBlank(encrypted)) {
-            return encryptionManager.decryptForInternalUse(encrypted);
         } else {
             return null;
         }
